create basic search with built in coffee search
