+++ frontend/src/components/IndexPage/index.js
@@ -6,7 +6,6 @@ import {
   clearErrors
 } from "../../store/businesses";
 import { useEffect } from "react";
-import { createSelector } from "@reduxjs/toolkit";
 import MainContent from "./MainContent";
 import TitleCard from "./TitleCard";
 
@@ -20,13 +19,7 @@ const IndexPage = () => {
   // const businessesObject = useSelector(getBusinesses);
   // const businesses = Object.values(businessesObject);
 
-  const readBusinesses = createSelector(
-    (state) => state.businesses,
-    (businesses) => {
-      return Object.values(businesses);
-    }
-  );
-  const businesses = useSelector(readBusinesses);
+  const businesses = useSelector(getBusinesses);
 
   // console.log(businesses);
   // useEffect(() => {
@@ -52,10 +45,10 @@ const IndexPage = () => {
     );
 
   return (
-    <>
+    <div id="index-page">
       <TitleCard businesses={businesses} />
       <MainContent businesses={businesses} />
-    </>
+    </div>
   );
 };
